Moves PriceAdjustments component into CartPage
